Special cast for gcc.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@384 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ net/src/IPAddress.cpp
@@ -20,13 +20,15 @@
 /*
  * vislib::net::IPAddress::ANY
  */
-const vislib::net::IPAddress vislib::net::IPAddress::ANY(INADDR_ANY);
+const vislib::net::IPAddress vislib::net::IPAddress::ANY(
+    static_cast<unsigned long>(INADDR_ANY));
 
 
 /*
  * vislib::net::IPAddress::NONE
  */
-const vislib::net::IPAddress vislib::net::IPAddress::NONE(INADDR_NONE);
+const vislib::net::IPAddress vislib::net::IPAddress::NONE(
+    static_cast<unsigned long>(INADDR_NONE));
 
 
 /*
